Hci model/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic